# Adds a pool of UTF8 String encoders