Most of the expense involved in constructing new binary writers
comes from allocating/initializing the buffers needed to encode
Java's UTF-16 Strings to UTF-8.

This change refactors the UTF-8 encoding logic into its own
class (Utf8StringEncoder) and introduces a singleton
Utf8StringEncoderPool that allows these encoders to be reused
across instantiations of binary writers.

### Benchmark

This test initializes a new binary writer, writes a small string (`"foo"`), then closes the writer in a tight loop. The source can be found [here](https://gist.github.com/zslayton/95730187f86f467ced87f13452ba9fb4).

**Before**
```
Benchmark                             Score           Error   Units
time                                 37.664 ±         3.661   ms/op
·gc.alloc.rate                      773.095 ±         3.882  MB/sec
·gc.alloc.rate.norm           438323910.400 ±    366718.519    B/op
·gc.churn.G1_Eden_Space             739.958 ±       238.206  MB/sec
·gc.churn.G1_Eden_Space.norm  419640115.200 ± 135676093.820    B/op
·gc.churn.G1_Old_Gen                  0.017 ±         0.033  MB/sec
·gc.churn.G1_Old_Gen.norm          9425.600 ±     18474.404    B/op
·gc.count                            20.000                  counts
·gc.time                             13.000                      ms
```

**After**
```
Benchmark                             Score           Error   Units
time                                 17.788 ±         3.631   ms/op
·gc.alloc.rate                       43.899 ±         0.850  MB/sec
·gc.alloc.rate.norm            24007104.000 ±    489259.500    B/op
·gc.churn.G1_Eden_Space              57.248 ±       182.466  MB/sec
·gc.churn.G1_Eden_Space.norm   31457280.000 ± 100263003.914    B/op
·gc.count                             2.000                  counts
·gc.time                              4.000                      ms
```

Following this change, the benchmark:
* Took 52.77% less time to run.
* Reduced its allocation rate from ~773MB/sec to ~44MB/sec (94.32% less)

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.