@tgregg 2021-06-23
Note: the binary reader also keeps per-instance buffers for string decoding, and could probably benefit similarly from using a pool. https://github.com/amzn/ion-java/blob/master/src/com/amazon/ion/impl/IonReaderBinaryRawX.java#L118