@zslayton 2021-06-23
@tgregg said:
> Can you also do a benchmark where instead of creating many writers that each write a single string, you create one writer that writes many strings? That way we can verify no impact to that use case.

Test data: a text Ion file with 10,000 repetitions of the top level string `"brevity is the soul of wit"` created with:
```
yes '"brevity is the soul of wit"' | head -n 10000 > /tmp/brevity.ion
```

`ion-java-benchmark-cli` command:
```
java -jar ion-java-benchmark-cli_VERSION.jar write --forks 2 /tmp/brevity.ion
```

**Before (v1.8.2)**
```
Benchmark                 Score      Error   Units
time                      2.418 ±    0.588   ms/op
:Heap usage               4.114 ±    0.005      MB
:Serialized size          0.280                 MB
:·gc.alloc.rate           0.106 ±    0.005  MB/sec
:·gc.alloc.rate.norm  56423.200 ± 2853.160    B/op
:·gc.count                  ≈ 0             counts
```

**After (this branch)**
```
Benchmark                 Score      Error   Units
time                      2.426 ±    0.458   ms/op
:Heap usage               4.156 ±    0.005      MB
:Serialized size          0.280                 MB
:·gc.alloc.rate           0.028 ±    0.005  MB/sec
:·gc.alloc.rate.norm  15007.200 ± 2810.004    B/op
:·gc.count                  ≈ 0             counts
```